clean up repo with better testing practices
