+++ tests/lib/rclone_test.py
@@ -1,0 +1,6 @@
+import os
+
+
+def test_rclone_installed():
+    """make sure you can run rclone version"""
+    assert os.system("rclone version") == 0
