+++ tests/readme.md
@@ -1,0 +1,1 @@
+All files in this directory are tests for scheduler. You should not import any code in this directory when developing in `userCode`/
