+++ userCode/lib/classes.py
@@ -11,7 +11,8 @@ from minio import Minio
 from urllib3 import BaseHTTPResponse
 from lakefs.client import Client
 
-from userCode.lib.lakefsUtils import create_branch_if_not_exists
+from userCode.lib.lakefs import LakeFSClient
+
 from .env import (
     GLEANER_MINIO_SECRET_KEY,
     GLEANER_MINIO_ACCESS_KEY,
@@ -130,6 +131,7 @@ class FileTransferer:
         path_to_file: str,
         destination_filename: str,
         destination_branch: str,
+        lakefs_client: LakeFSClient,
     ):
         """
         Copy a file from minio to lakefs
@@ -142,7 +144,7 @@ class FileTransferer:
 
         get_dagster_logger().info(f"Uploading {path_to_file} to {LAKEFS_ENDPOINT_URL}")
 
-        new_branch = create_branch_if_not_exists(destination_branch)
+        new_branch = lakefs_client.create_branch_if_not_exists(destination_branch)
 
         self._run_subprocess(
             f"rclone copyto minio:{GLEANER_MINIO_BUCKET}/{path_to_file} lakefs:geoconnex/{destination_branch}/{destination_filename} -v"
