--- userCode/test/__init__.py
@@ -1,3 +1,0 @@
-# Copyright 2025 Lincoln Institute of Land Policy
-# SPDX-License-Identifier: Apache-2.0
-
